bagitobjecttransfer/requirements.txt
git+https://github.com/whikloj/django-dbtemplates@3.0.3#egg=django-dbtemplates
git+https://github.com/Azure-Samples/ms-identity-python-samples-common@0.16.6#egg=ms_identity_web
asgiref==3.6
bagit==1.8.1
clamd==1.0.2
click==8.0.3
colorama==0.4.4
dateparser==1.1.1
Deprecated==1.2.13
Django==4.2
django-countries==7.2.1
django-formtools==2.3
django-override-storage==0.3.2
django-recaptcha==3.0.0
django-rq==2.8.0
gunicorn==20.1.0
importlib-metadata==4.8.3
msal==1.22.0
mysql-connector-python==8.0.24
packaging==21.3
protobuf==3.19.3
PyMySQL==1.0.2
pyparsing==3.0.6
python-decouple==3.5
pytz==2021.3
redis==4.1.0
rq==1.10.1
sqlparse==0.4.2
typing_extensions==4.0.1
wrapt==1.13.3
zipp==3.6.0
